[BE] (feat) findAll() 메서드 수정
[BE] (feat) findAll() 메서드 수정

테스트 수정
