+++ 4 WEEK/item-service/src/main/java/hello/itemservice/service/ItemService.java
@@ -2,6 +2,8 @@ package hello.itemservice.service;
 
 import hello.itemservice.domain.Item;
 
+import java.util.List;
+
 public interface ItemService {
 
     //상품 목록 추가
@@ -9,4 +11,8 @@ public interface ItemService {
 
     //상품상세
     Item findItem(Long itemId);
+
+    List<Item> findAll();
 }
+
+
